# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaaS ERP API",
        "description": "API Documentation for the Multi-tenant SaaS ERP Application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/v1/tenants": {
            "get": {
                "tags": [
                    "Tenants"
                ],
                "summary": "List all tenants",
                "description": "Display a listing of tenants.",
                "operationId": "4e18e9d4532e971597f302e1fb79ea07",
                "responses": {
                    "200": {
                        "description": "List of tenants retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "tenants": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Create a new tenant",
                "description": "Store a newly created tenant.",
                "operationId": "077fa4ae7fd0495ce75cfe55ebd761e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug",
                                    "owner_first_name",
                                    "owner_last_name",
                                    "owner_email",
                                    "owner_password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Acme Corp"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "acme-corp"
                                    },
                                    "domain": {
                                        "type": "string",
                                        "example": "acme.saas.com",
                                        "nullable": true
                                    },
                                    "owner_first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "owner_last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "owner_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "owner_password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tenant created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "tenant": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tenants/{tenant}": {
            "get": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Get specific tenant details",
                "description": "Display the specified tenant.",
                "operationId": "81579798d81c922b478baa8946b1e999",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "Tenant UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant details retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "tenant": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Update specific tenant",
                "description": "Update the specified tenant.",
                "operationId": "20469ad3bde6961ae665652fb9173ca5",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "Tenant UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "domain": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tenant updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "tenant": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Delete specific tenant",
                "description": "Remove the specified tenant.",
                "operationId": "81d8ddaa75f3201b7ad16b8a3afe32e0",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "Tenant UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}